Bugfix columns were not displaying properly when editing weeks
When editing a week that had missing assignments, the empty cell and corresponding select list systematically ended up in the left column,
regardless of what roles they were for. So all select lists were in the wrong column - infrastructure people in the
oncall_weekday_developer column, etc.

This fixes it, however it adds an enumeration over roles, and two if statements, which I feel are
slowing down the app. Editing a week and creating one used to be faster.